Consolidate C# and VB language servers into one
