+++ eng/targets/GenerateServiceHubConfigurationFiles.targets
@@ -10,8 +10,7 @@
     <ServiceHubService Include="roslynRemoteHost" ClassName="Microsoft.CodeAnalysis.Remote.RemoteHostService" />
     <ServiceHubService Include="roslynSnapshot" ClassName="Microsoft.CodeAnalysis.Remote.SnapshotService" />
     <ServiceHubService Include="roslynRemoteSymbolSearchUpdateEngine" ClassName="Microsoft.CodeAnalysis.Remote.RemoteSymbolSearchUpdateEngine" />
-    <ServiceHubService Include="roslynCSharpLanguageServer" ClassName="Microsoft.CodeAnalysis.Remote.CSharpLanguageServer" />
-    <ServiceHubService Include="roslynVisualBasicLanguageServer" ClassName="Microsoft.CodeAnalysis.Remote.VisualBasicLanguageServer" />
+    <ServiceHubService Include="roslynLanguageServer" ClassName="Microsoft.CodeAnalysis.Remote.LanguageServer" />
   </ItemGroup>
 
   <PropertyGroup>
