+++ src/VisualStudio/CSharp/Impl/xlf/CSharpVSResources.cs.xlf
@@ -22,11 +22,6 @@
         <target state="translated">Pravidla formátování jazyka C#</target>
         <note />
       </trans-unit>
-      <trans-unit id="CSharp_language_server_client">
-        <source>C# Language Server Client</source>
-        <target state="translated">Klient serveru pro jazyk C#</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Discard">
         <source>Discard</source>
         <target state="translated">Zahodit</target>
