+++ src/VisualStudio/CSharp/Impl/xlf/CSharpVSResources.es.xlf
@@ -22,11 +22,6 @@
         <target state="translated">Reglas de formato de C#</target>
         <note />
       </trans-unit>
-      <trans-unit id="CSharp_language_server_client">
-        <source>C# Language Server Client</source>
-        <target state="translated">Cliente de servidor del lenguaje C#</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Discard">
         <source>Discard</source>
         <target state="translated">Descartar</target>
