+++ src/VisualStudio/CSharp/Impl/xlf/CSharpVSResources.ko.xlf
@@ -22,11 +22,6 @@
         <target state="translated">C# 서식 설정 규칙</target>
         <note />
       </trans-unit>
-      <trans-unit id="CSharp_language_server_client">
-        <source>C# Language Server Client</source>
-        <target state="translated">C# 언어 서버 클라이언트</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Discard">
         <source>Discard</source>
         <target state="translated">폐기</target>
