+++ src/VisualStudio/CSharp/Impl/xlf/CSharpVSResources.tr.xlf
@@ -22,11 +22,6 @@
         <target state="translated">C# biçimlendirme kuralları</target>
         <note />
       </trans-unit>
-      <trans-unit id="CSharp_language_server_client">
-        <source>C# Language Server Client</source>
-        <target state="translated">C# Dil Sunucusu İstemcisi</target>
-        <note />
-      </trans-unit>
       <trans-unit id="Discard">
         <source>Discard</source>
         <target state="translated">At</target>
