+++ src/VisualStudio/Core/Def/xlf/ServicesVSResources.ko.xlf
@@ -72,6 +72,11 @@
         <target state="translated">빌드 + 실시간 분석(NuGet 패키지)</target>
         <note />
       </trans-unit>
+      <trans-unit id="CSharp_VB_Language_Server_Client">
+        <source>C#/Visual Basic Language Server Client</source>
+        <target state="new">C#/Visual Basic Language Server Client</target>
+        <note />
+      </trans-unit>
       <trans-unit id="Calculating_dependents">
         <source>Calculating dependents...</source>
         <target state="translated">종속 항목을 계산하는 중...</target>
