+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.de.xlf
@@ -277,11 +277,6 @@
         <target state="translated">VB-Codierungskonvention</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="translated">Client des Visual Basic-Sprachservers</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'Überprüfung auf "nothing":</target>
