+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.es.xlf
@@ -277,11 +277,6 @@
         <target state="translated">Convenciones de código de VB</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="translated">Cliente de servidor del lenguaje Visual Basic</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'Comprobación de "nothing":</target>
