+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.fr.xlf
@@ -277,11 +277,6 @@
         <target state="translated">Conventions de codage VB</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="translated">Client du serveur de langage Visual Basic</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'Contrôle de 'nothing' :</target>
