+++ src/VisualStudio/VisualBasic/Impl/xlf/BasicVSResources.tr.xlf
@@ -277,11 +277,6 @@
         <target state="translated">VB kodlama kuralları</target>
         <note />
       </trans-unit>
-      <trans-unit id="Visual_Basic_language_server_client">
-        <source>Visual Basic Language Server Client</source>
-        <target state="translated">Visual Basic Dil Sunucusu İstemcisi</target>
-        <note />
-      </trans-unit>
       <trans-unit id="nothing_checking_colon">
         <source>'nothing' checking:</source>
         <target state="translated">'nothing' denetimi:</target>
